Change: Set width and height in frame parameters
At least for me, this gives the same results as using
set-frame-height/width, but without the conflict that requires
sleep-for.  The docs say that it sets the width/height of the frame's
text area in pixels, which I guess could be different if non-text-area
features like scroll bars, menu bars, or tool bars are enabled.  But I
think this is good enough, and it fixes the silly sleep-for problem.