Add: Track focus
This way, if the Yequake frame is open, but doesn't have input focus
in the window system, we can know, and raise/focus the frame again
instead of deleting it.

As far as I can tell, this is the only way to do this, because there
isn't a function like frame-focused-p.