Add: left and top frame parameters
It would be preferable to pass them directly as frame parameters, but
on my system, it didn't work consistently: when using float ratios, it
seemed that if another window overlapped the position the Yequake
frame should appear in, it would appear offset instead of in the right
place.  By converting float ratios to integer values ourselves, it
seems to always work correctly.

Fixes #1.  Thanks to Ilya Khaprov (@deadtrickster) for suggesting.