Add: Set user-position
According to the elisp manual, it should help ensure the window system
puts the frame where we want it, but it's not guaranteed that it will
be honored.  Indeed, KWin from KDE 4 seems to ignore it.  But we might
as well add it, because it might make the difference for someone.