Change: Set frame name once and pass to --toggle
It's not very useful to specify the window system frame name
separately from the Yequake configuration frame name.  They're just
strings, so one name is enough.

Also, this makes it possible to retoggle a frame by name even if the
frame isn't currently visible.  This seems like the simplest way to
make this change.

NOTE: This commit introduces a minor bug breaking retoggling, which
will be fixed in the next commit which fixes yequake-org-capture.
It's not worth the trouble to make this commit stand completely alone.